fix: extract pure preference logic from Firebase-coupled firestore.ts
lib/firestore.ts imports lib/firebase.ts, which calls initializeApp()/
getAuth() at module load, so importing the file just to exercise the
pure normalizePreferences() helper booted live Firebase and threw
auth/invalid-api-key without env vars. The prior fix patched npm test
to load .env.local, which masked the coupling for the npm script but
still failed running a single test file directly, and made the suite
silently depend on env config.

Move FavoriteService, UserPreferences, normalizePreferences(), and
isFavoriteService() into a new Firebase-free lib/preferences.ts.
lib/firestore.ts re-exports them so existing imports (lib/AuthContext.tsx)
keep working, while loadUserPreferences/saveUserPreferences stay put.
Revert package.json's test script to the plain, quoted glob form.
Repoint tests/userPreferences.test.ts at lib/preferences and add a case
proving isFavoriteService now rejects a NaN id via Number.isFinite
instead of typeof === "number".
