+++ lib/preferences.ts
@@ -1,0 +1,46 @@
+// lib/preferences.ts
+//
+// Pure data-shaping logic for user preferences, deliberately kept free of any
+// Firebase import. lib/firestore.ts imports lib/firebase.ts, which calls
+// initializeApp()/getAuth() at module load — importing that module just to
+// test normalizePreferences() boots live Firebase and throws without env
+// vars. Keeping this file dependency-free lets it be unit tested standalone.
+
+export type FavoriteService = {
+  id: number;
+  name: string;
+  logoPath: string;
+};
+
+export type UserPreferences = {
+  favoriteCountries: string[];
+  darkMode: boolean;
+  favoriteServices: FavoriteService[];
+};
+
+// Firestore documents are untrusted input: a partial write or an older client
+// can leave any field missing or the wrong shape. Normalizing in one place keeps
+// the defaults identical for the loader and for callers rebuilding local state.
+export function normalizePreferences(data: unknown): UserPreferences {
+  const record = (data ?? {}) as Record<string, unknown>;
+
+  const favoriteServices = Array.isArray(record.favoriteServices)
+    ? record.favoriteServices.filter(isFavoriteService).map((service) => ({
+        id: service.id,
+        name: service.name,
+        logoPath: typeof service.logoPath === "string" ? service.logoPath : "",
+      }))
+    : [];
+
+  return {
+    favoriteCountries: Array.isArray(record.favoriteCountries) ? record.favoriteCountries : [],
+    darkMode: typeof record.darkMode === "boolean" ? record.darkMode : true,
+    favoriteServices,
+  };
+}
+
+function isFavoriteService(value: unknown): value is { id: number; name: string; logoPath?: unknown } {
+  if (typeof value !== "object" || value === null) return false;
+  const candidate = value as Record<string, unknown>;
+  return Number.isFinite(candidate.id) && typeof candidate.name === "string";
+}
