+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test": "node --env-file-if-exists=.env.local --import tsx --test 'tests/**/*.test.ts'"
+    "test": "node --import tsx --test 'tests/**/*.test.ts'"
   },
   "dependencies": {
     "@heroicons/react": "^2.2.0",
